geoatacama/settings.py
import os
from pathlib import Path
import environ
from datetime import timedelta


# Build paths inside the project like this: BASE_DIR / 'subdir'.
BASE_DIR = Path(__file__).resolve().parent.parent
TEMPLATES_DIR = os.path.join(BASE_DIR,'templates')
STATIC_DIR = os.path.join(BASE_DIR,'static')

env = environ.Env()
environ.Env.read_env(os.path.join(BASE_DIR, '.env'))

# SECURITY WARNING: keep the secret key used in production secret!
SECRET_KEY = 'django-insecure-29k78=1bk13ayqhwb*p&0z3j!*)%69w$%7$$n!al$&q-m6%=am'

# SECURITY WARNING: don't run with debug turned on in production!
DEBUG = True

ALLOWED_HOSTS = ["*"]

CSRF_TRUSTED_ORIGINS = ['https://www.sicgeoatacama.cl','https://sicgeoatacama.cl',"http://localhost","http://127.0.0.1"]

MESSAGE_STORAGE = "django.contrib.messages.storage.cookie.CookieStorage"

# Application definition
INSTALLED_APPS = [
    'django.contrib.admin',
    'django.contrib.auth',
    'django.contrib.contenttypes',
    'django.contrib.sessions',
    'django.contrib.messages',
    'django.contrib.staticfiles',
    'widget_tweaks',
    'crispy_forms',
    'crispy_bootstrap4',
    'core',
    'user',
    'vehicle',
    'mining',
    'maintenance',
    'machine',
    'documentation',
    'messenger',
    'planning',
    'drilling',
    'checklist',
    'inventory',
    'equipment',
    'session_security',
    'rest_framework',
    'corsheaders',
]

CRISPY_TEMPLATE_PACK = 'bootstrap4'

MIDDLEWARE = [
    'django.middleware.security.SecurityMiddleware',
    'django.contrib.sessions.middleware.SessionMiddleware',
    'django.middleware.common.CommonMiddleware',
    'django.middleware.csrf.CsrfViewMiddleware',
    'django.contrib.auth.middleware.AuthenticationMiddleware',
    'django.contrib.messages.middleware.MessageMiddleware',
    'django.middleware.clickjacking.XFrameOptionsMiddleware',
    'corsheaders.middleware.CorsMiddleware',
]

ROOT_URLCONF = 'geoatacama.urls'

TEMPLATES = [
    {
        'BACKEND': 'django.template.backends.django.DjangoTemplates',
        'DIRS': [TEMPLATES_DIR],
        'APP_DIRS': True,
        'OPTIONS': {
            'context_processors': [
                'django.template.context_processors.debug',
                'django.template.context_processors.request',
                'django.contrib.auth.context_processors.auth',
                'django.contrib.messages.context_processors.messages',
            ],
            'libraries': {
                'custom_filters': 'checklist.templatetags.custom_filters',
            },
        },
    },
]

WSGI_APPLICATION = 'geoatacama.wsgi.application'


# Database
# https://docs.djangoproject.com/en/5.0/ref/settings/#databases

DATABASES = {
    'default': {
        'ENGINE': 'django.db.backends.mysql', 
        'NAME': env('SQL_NAME'),
        'PORT': env('SQL_PORT'),
        'USER': env('SQL_USER'),  
        'PASSWORD': env('SQL_PASSWORD'),
        'OPTIONS': {
            'init_command': env('DB_OPTIONS_INIT_COMMAND'),
            'charset': env('DB_OPTIONS_CHARSET'),
        },
    }
}


# Password validation
# https://docs.djangoproject.com/en/5.0/ref/settings/#auth-password-validators

AUTH_PASSWORD_VALIDATORS = [
    {
        'NAME': 'django.contrib.auth.password_validation.UserAttributeSimilarityValidator',
    },
    {
        'NAME': 'django.contrib.auth.password_validation.MinimumLengthValidator',
    },
    {
        'NAME': 'django.contrib.auth.password_validation.CommonPasswordValidator',
    },
    {
        'NAME': 'django.contrib.auth.password_validation.NumericPasswordValidator',
    },
]


# Internationalization
# https://docs.djangoproject.com/en/5.0/topics/i18n/

LANGUAGE_CODE = 'es-cl'

TIME_ZONE = 'America/Santiago'

USE_I18N = True

USE_TZ = True


# Static files (CSS, JavaScript, Images)
# https://docs.djangoproject.com/en/5.0/howto/static-files/

STATIC_URL = 'static/'
STATICFILES_DIRS = [STATIC_DIR]
STATIC_ROOT = os.path.join(BASE_DIR, 'staticfiles')

MEDIA_URL = '/media/'
MEDIA_ROOT = os.path.join(BASE_DIR,'media')


# Default primary key field type
# https://docs.djangoproject.com/en/5.0/ref/settings/#default-auto-field

DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'

AUTH_USER_MODEL = "user.User"

LOGIN_REDIRECT_URL = "select"
LOGOUT_REDIRECT_URL = "logincustom"
LOGIN_URL = "logincustom"


#SESSION_SAVE_EVERY_REQUEST = True
#SESSION_COOKIE_AGE = 60 * 20
SESSION_EXPIRE_AT_BROWSER_CLOSE = True

#SESSION_SECURITY_WARN_AFTER = 480 #envia un aviso
#SESSION_SECURITY_EXPIRE_AFTER = 600 #cierra la sesion
if os.name == 'posix':  # Linux y macOS
    logs_dir = '/home/app/Geoatacama/geoatacama/logs/'
elif os.name == 'nt':   # Windows
    logs_dir = os.path.join(BASE_DIR, 'log_django')
LOGGING = {
    'version': 1,
    'disable_existing_loggers': False,
    'formatters': {
        'verbose': {
            'format': '{levelname} {asctime} {module} {message}',
            'style': '{',
        },
        'simple': {
            'format': '{levelname} {message}',
            'style': '{',
        },
    },
    'handlers': {
        'file': {
            'level': 'ERROR',
            'class': 'logging.FileHandler',
            'filename': os.path.join(logs_dir, 'django_error.log'),
            'formatter': 'verbose',
        },
    },
    'loggers': {
        'django': {
            'handlers': ['file'],
            'level': 'ERROR',
            'propagate': True,
        },
    },
}

TWILIO_ACCOUNT_SID = env('TWILIO_ACCOUNT_SID')
TWILIO_AUTH_TOKEN = env('TWILIO_AUTH_TOKEN')
TWILIO_WHATSAPP_NUMBER = env('TWILIO_WHATSAPP_NUMBER')
TWILIO_PHONE_NUMBER = env('TWILIO_PHONE_NUMBER')

DEFAULT_FROM_EMAIL = env('DEFAULT_FROM_EMAIL')
EMAIL_BACKEND = 'django.core.mail.backends.smtp.EmailBackend'
EMAIL_HOST = 'smtp.office365.com'   
EMAIL_PORT = 587
EMAIL_USE_TLS = True
EMAIL_HOST_USER = DEFAULT_FROM_EMAIL
EMAIL_HOST_PASSWORD = env('EMAIL_HOST_PASSWORD')

CELERY_BROKER_URL = 'redis://localhost:6379/0'
CELERY_RESULT_BACKEND = 'redis://localhost:6379/0'
CELERY_ACCEPT_CONTENT = ['application/json']
CELERY_TASK_SERIALIZER = 'json'
CELERY_RESULT_SERIALIZER = 'json'
CELERY_TIMEZONE = 'America/Santiago'

REST_FRAMEWORK = {
    'DEFAULT_AUTHENTICATION_CLASSES': (
        'rest_framework_simplejwt.authentication.JWTAuthentication',
    ),
}

SIMPLE_JWT = {
    # Duración del token de acceso (aumenta según necesidad)
    #'ACCESS_TOKEN_LIFETIME': timedelta(minutes=60),  # Cambia a 60 minutos
    # Duración del token de refresh (aumenta según necesidad)
    'REFRESH_TOKEN_LIFETIME': timedelta(days=18),     # Cambia a 7 días
    "TOKEN_OBTAIN_SERIALIZER": "api.serializers.CustomTokenObtainPairSerializer",
    # Opcional: Puedes evitar que un token de refresh caduque permanentemente
    #'ROTATE_REFRESH_TOKENS': True,  # Gira el token de refresh después de usarlo
    #'BLACKLIST_AFTER_ROTATION': True,  # Añade los tokens antiguos a una lista negra
}

CORS_ALLOW_ALL_ORIGINS = True
CORS_ALLOW_CREDENTIALS = True

CORS_ALLOWED_ORIGINS = [
    "http://localhost:3000",  # React o cualquier cliente local
    "https://sicgeoatacama.cl",
    "https://www.sicgeoatacama.cl",
    "http://127.0.0.1:8000",
    "http://localhost:8081",
]

CORS_ALLOW_HEADERS = [
    'content-type',
    'authorization',
    'x-csrftoken',
]

CORS_ALLOW_METHODS = [
    'GET',
    'POST',
    'PUT',
    'PATCH',
    'DELETE',
    'OPTIONS',
]

if DEBUG:
    BASE_API_URL = "http://127.0.0.1:8000"
else:
    BASE_API_URL = "https://sicgeoatacama.cl"

# Construimos las URLs dinámicamente
DASHBOARD_INVENTARIO_PREVENCION_URL = f'{BASE_API_URL}/api/dashboardInventarioPrevencion/?format=json'
DASHBOARD_INVENTARIO_SONDAJES_URL = f'{BASE_API_URL}/api/dashboardInventarioSondaje/?format=json'
DASHBOARD_INVENTARIO_VEHICULOS_URL = f'{BASE_API_URL}/api/dashboardInventarioVehiculo/?format=json'
DASHBOARD_VEHICULOS_URL = f'{BASE_API_URL}/api/dashboardVehiculos/?format=json'