requirements.txt
accessible-pygments==0.0.4
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiohttp-retry==2.8.3
aiosignal==1.3.1
alabaster==1.0.0
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
billiard==4.2.1
cachetools==5.5.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap4==2024.1
cryptography==43.0.1
cssselect2==0.7.0
Django==4.2.11
django-appconf==1.0.6
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-environ==0.11.2
django-pwa==2.0.1
django-select2==8.2.1
django-session-security==2.6.7
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.20.1
et_xmlfile==2.0.0
frozenlist==1.4.1
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.67.0
html5lib==1.1
httplib2==0.22.0
idna==3.10
imagesize==1.4.1
Jinja2==3.1.5
kombu==5.4.2
lxml==5.3.0
MarkupSafe==2.1.5
multidict==6.1.0
mysql-connector-python==9.5.0
mysqlclient==2.2.4
numpy==2.2.4
oauthlib==3.2.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==10.4.0
prompt_toolkit==3.0.48
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydata-sphinx-theme==0.15.4
Pygments==2.19.1
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
PyMySQL==1.1.1
pyngrok==7.2.3
pyparsing==3.1.4
pypdf==5.0.1
pypng==0.20220715.0
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.4
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
rut-chile==2.0.1
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.1.3
sphinx-autodoc-typehints==3.0.1
sphinx-book-theme==1.1.4
sphinx-copybutton==0.5.2
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.3.0
tomli==2.2.1
twilio==9.3.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.16
yarl==1.13.1
